pyproject.toml
[tool.poetry]
name = "simple-api"
version = "0.1.0"
description = "a simple api to handle user shifts"
authors = ["Khalid Al-Awady <pipes_and_bottles@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy-Utils = "^0.37.8"
psycopg2-binary = "^2.9.1"
bcrypt = "^3.2.0"
alembic = "^1.7.6"
faker = "^26.0.0"
httpx = "^0.27.0"
fastapi = "^0.111.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
mypy = "^0.910"
pdbpp = "^0.10.3"
ipython = "^7.26.0"
black = "^21.7b0"
flake8 = "^3.9.2"
requests = "^2.26.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
